+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added the ability to change channel message display mode (chat, lfrp, both)
 - Added an 'about' panel to the menu, to always have access to developer info.
 - Removed a lot of CSS transitions for performance, especially on mobile.
+- Removed info footer in favor of a link with application information
 
 # 0.2.1
 - Fixed a nasty bug where the login menu didn't show up.