+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Removed info footer in favor of a link with application information
 - Log out option works
 - Switch character option works
+- The "Remember me" checkbox works (rather, stops working when you leave it unchecked)
 
 # 0.2.1
 - Fixed a nasty bug where the login menu didn't show up.